+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Breaking changes
 
+- Remove `attrs["crs"]` usage from `GeoVectorDataset` and child classes (`Flight`, `Fleet`). All spatial data is assumed to be EPSG:4326 (WGS84). This was previously assumed implicitly, but now the `crs` attribute is removed from the `attrs` dictionary.
 - Remove deprecated `MetDataset.variables` property in favor of `MetDataset.indexes`.
 - Remove `**kwargs` in `MetDataArray` constructor.
 - Rename `ARCOERA5` to `ERA5ARCO` for consistency with the `ERA5` and `ERA5ModelLevel` interfaces.